Django_React/DjangoBlogProject/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
mysqlclient>=2.0,<3.0